Refactor _h5helpers into types.base
